+++ src/TableColumn/Link.php
@@ -1,17 +1,31 @@
 <?php
 
-namespace atk4\ui\Column;
+namespace atk4\ui\TableColumn;
 
 /**
  * Implements Column helper for grid.
  */
 class Link extends Generic
 {
+    /**
+     * TODO: Refactor so that this goes away!!
+     */
     public $page = null;
 
+    /**
+     * Use
+     *   new Link('http://yahoo.com?id={$id}');
+     * or
+     *   new Link(['order', 'id'=>'{$id}']);.
+     *
+     * @param string|array $page Destination definition
+     */
     public function __construct($page = [])
     {
-        $this->page = $page[0];
+        if (!is_array($page)) {
+            $page = [$page];
+        }
+        $this->page = $page;
     }
 
     /**
@@ -19,9 +33,6 @@ class Link extends Generic
      */
     public function getCellTemplate(\atk4\data\Field $f)
     {
-        if (!is_array($this->page)) {
-            return $this->app->getTag('td', [], $this->app->getTag('a', ['href'=>$this->page], '{$'.$f->short_name.'}'));
-        }
         foreach ($this->page as &$val) {
             $val = str_replace('{$', '___o', $val);
             $val = str_replace('}', 'c___', $val);